refactor: Vite config, findAllHtmlFiles 리팩토링
